Delete .github/workflows/dependabot-approve.yml (#13)
We still require human approval. But maybe that could change!

<!--

Thank you for contributing to the watchtower project! 🙏

We truly appreciate all the contributions we get from the community.

To make your PR experience as smooth as possible, make sure that you
include the following in your PR:

- What your PR contributes
- Which issues it solves (preferrably using auto closing instructions
like "closes #123".
- Tests that verify the code your contributing
- Updates to the documentation

Thank you again! ✨

-->